crypto/hmac: Fix panic in unreachable code in test
Instead of unconditionally panicing we should make an exception for
tests, as other parts of the code does. This still accomplishes the
original goal of not allowing non-FIPS code to execute, but also allows
the stdlib tests to run without causing their own panic from a nil
pointer deref because of FIPS mode not being able to work with a
unapproved hash.