Specify explicit dependency versions (#64)
Currently the scripts target the latest commit on a given branch.
This is because the original intent was to set up a source branch
once and then continually rebase upstream Go on top of it.  However,
this is not the way the workflow has played out so far.  For
Go 1.19+, we're currently distributing as scripts and patches
that generate a golang-fips tree on top of an upstream Go ref.

This commit introduces a versions.json file with a mapping between
repo versions and git refs for golang/go, golang-fips/go, and
golang-fips/openssl-fips.  The setup scripts now read from
versions.json when selecting the default git refs, allowing
us to standardize dependencies to known working versions
where a given commit of golang-fips/go will behave consistently.