Remove the call to parallelize TestEverything
This method, while good for performance, is interfering with OpenSSL and
Go interactions by accidentally hiding errors. This is due to likely to
the goroutine being preempted during CGO calls. There is a series of
circumstances that leads to the OpenSSL thread local error state being
overwritten causing us to lose the error, and thus fail the test.